Moved from float to uint32_t for TX power, substituted dBm with mW in CPS channel struct
General code optimization allowing to reduce the overall binary size, especially
for devices without hardware floating point support. On the MD-380 (which has the
FPU) the .text size has been reduced of 4'928 bytes.

+++ openrtx/include/rtx/rtx.h
@@ -43,7 +43,7 @@ typedef struct
     freq_t rxFrequency;     /**< RX frequency, in Hz           */
     freq_t txFrequency;     /**< TX frequency, in Hz           */
 
-    float    txPower;       /**< TX power, in W                */
+    uint32_t txPower;       /**< TX power, in mW               */
     uint8_t  sqlLevel;      /**< Squelch opening level         */
 
     uint16_t rxToneEn : 1,  /**< RX CTC/DCS tone enable        */
